Initial commit for C++ DLL Class Interface compiling for VS2017

+++ pslr_enum.h
@@ -181,7 +181,7 @@ typedef enum {
     PSLR_SCENE_MODE_MAX
 } pslr_scene_mode_t;
 
-int str_comparison_i (const char *s1, const char *s2, int n);
+int str_comparison_i (const char *s1, const char *s2, size_t n);
 int find_in_array( const char** array, int length, char* str );
 
 pslr_color_space_t get_pslr_color_space( char *str );
